Move IActiveConfigScopeTracker inside Core assembly.
Due to the fact that it needs to be accesible to the Infrastructure.VS.
Part of SLVS-1497
